+++ ChangeLog.md
@@ -52,4 +52,5 @@
  * As an additional option to the previous change, players can opt for a defensively resistant opponent. For instance, Magnemite would normally face a Ground type, but this option makes the Gym use an Electric type as Electric resists both Electric and Steel.
  * Modified the following sections of in-game text to match choices
     * Striaton City Gym leaders will use their true types when using type-theming on Gym Leaders, or use generic terms when completely random
+   * Nacrene City Gym Guy will use true type and appropriate weakness when using type-theming on Gym Leaders, or use generic terms when completely random
  * Added `Normalize` and `Stall` to list of negative abilities that can be banned